Add Windows (MSVC) support
Major changes:
- CMakeLists.txt: Windows platform detection, vcpkg integration, Winsock linking
- platform.h: Cross-platform abstractions for sockets, threading, file ops
- thpool.cc: C++ version with std::thread/mutex/condition_variable
- exec.cc: Windows CreateProcess implementation with PATHEXT support
- server.cc: SetConsoleCtrlHandler for signals, Windows random seeding
- network.cc: WSAStartup/WSACleanup, ioctlsocket for non-blocking
- background.cc: Socket pair instead of pipe (Windows select limitation)
- fileio.cc: Windows directory ops, backslash path handling

Build tested with MSVC 2022, ~95% of Ruby test suite passing.
Remaining test failures are platform differences (CRLF, env vars).

+++ src/db_verbs.cc
@@ -21,6 +21,7 @@
 
 #include <assert.h>
 #include <ctype.h>
+#include <stdint.h>
 
 #include <stdlib.h>
 #include <string.h>
@@ -577,7 +578,7 @@ try_again:
 
         assert(o != nullptr);
 
-        unsigned long first_parent_with_verbs = (unsigned long)o;
+        uintptr_t first_parent_with_verbs = (uintptr_t)o;
 
         /* found something with verbdefs, now check the cache */
         unsigned int hash, bucket;
